feat: update eth/tasks/admin.ts

+++ eth/contracts/facets/DFWhitelistFacet.sol
@@ -22,15 +22,15 @@ contract DFWhitelistFacet is WithStorage {
         return ws().allowedAccountsArray.length;
     }
 
-    function enableWhitelist() public onlyAdmin{
+    function enableWhitelist() public onlyAdmin {
         ws().enabled = true;
     }
 
-    function disableWhitelist() public onlyAdmin{
+    function disableWhitelist() public onlyAdmin {
         ws().enabled = false;
     }
 
-    function getWhitelistEnabled() public view returns(bool){
+    function getWhitelistEnabled() public view returns (bool) {
         return ws().enabled;
     }
 
@@ -103,11 +103,11 @@ contract DFWhitelistFacet is WithStorage {
         ws().allowedAccountsArray.push(toAdd);
     }
 
-    function batchAddToWhitelist(address[] memory addrs) public onlyAdmin{
-        for(uint i = 0;i<addrs.length;i++){
+    function batchAddToWhitelist(address[] memory addrs) public onlyAdmin {
+        for (uint256 i = 0; i < addrs.length; i++) {
             address toAdd = addrs[i];
-            if(!ws().allowedAccounts[toAdd]){
-                ws().allowedAccounts[toAdd]=true;
+            if (!ws().allowedAccounts[toAdd]) {
+                ws().allowedAccounts[toAdd] = true;
                 ws().allowedAccountsArray.push(toAdd);
             }
         }
